Fix: move state init to a service
